+++ MBCalendarKit/CalendarKit/Categories/Foundation/NSDate/NSDate+Description.m
@@ -13,7 +13,7 @@
 
 - (NSString *)description
 {
-    NSDateFormatter *formatter = CKCache.sharedCache.dateFormatter;
+    NSDateFormatter *formatter = [CKCache.sharedCache dateFormatterWithFormat:@""];
     formatter.timeStyle = NSDateFormatterNoStyle;
     formatter.dateStyle = NSDateFormatterLongStyle;
     return [formatter stringFromDate:self];
@@ -21,84 +21,75 @@
 
 - (NSString *)dayNameOnCalendar:(NSCalendar *)calendar
 {
-    NSDateFormatter *f = CKCache.sharedCache.dateFormatter;
+    NSDateFormatter *f = [CKCache.sharedCache dateFormatterWithFormat:@"ccc"];
     f.calendar = calendar;
     f.locale = calendar.locale;
-    [f  setLocalizedDateFormatFromTemplate: @"ccc"];
     return [f stringFromDate:self];
 }
 
 - (NSString *)monthNameOnCalendar:(NSCalendar *)calendar
 {
-    NSDateFormatter *f = CKCache.sharedCache.dateFormatter;
+    NSDateFormatter *f = [CKCache.sharedCache dateFormatterWithFormat:@"MMMM"];
     f.calendar = calendar;
     f.locale = calendar.locale;
-    [f  setLocalizedDateFormatFromTemplate:@"MMMM"];
     return [f stringFromDate:self];
 }
 
 - (NSString *)monthAndYearOnCalendar:(NSCalendar *)calendar
 {
-    NSDateFormatter *f = CKCache.sharedCache.dateFormatter;
+    NSDateFormatter *f = [CKCache.sharedCache dateFormatterWithFormat:@"MMMM yyyy"];
     f.calendar = calendar;
     f.locale = calendar.locale;
-    [f  setLocalizedDateFormatFromTemplate:@"MMMM yyyy"];
     return [f stringFromDate:self];
 }
 
 - (NSString *)monthAbbreviationAndYearOnCalendar:(NSCalendar *)calendar
 {
-    NSDateFormatter *f = CKCache.sharedCache.dateFormatter;
+    NSDateFormatter *f = [CKCache.sharedCache dateFormatterWithFormat:@"MMM yyyy"];
     f.calendar = calendar;
     f.locale = calendar.locale;
-    [f  setLocalizedDateFormatFromTemplate:@"MMM yyyy"];
     return [f stringFromDate:self];
 }
 
 
 - (NSString *)monthAbbreviationOnCalendar:(NSCalendar *)calendar
 {
-    NSDateFormatter *f = CKCache.sharedCache.dateFormatter;
+    NSDateFormatter *f = [CKCache.sharedCache dateFormatterWithFormat:@"MMM"];
     f.calendar = calendar;
     f.locale = calendar.locale;
-    [f  setLocalizedDateFormatFromTemplate:@"MMM"];
     return [f stringFromDate:self];
 }
 
 - (NSString *)monthAndDayOnCalendar:(NSCalendar *)calendar
 {
-    NSDateFormatter *f = CKCache.sharedCache.dateFormatter;
+    NSDateFormatter *f = [CKCache.sharedCache dateFormatterWithFormat:@"MMM d"];
     f.calendar = calendar;
     f.locale = calendar.locale;
-    [f  setLocalizedDateFormatFromTemplate:@"MMM d"];
     return [f stringFromDate:self];
 }
 
 - (NSString *)dayOfMonthOnCalendar:(NSCalendar *)calendar
 {
-    NSDateFormatter *f = CKCache.sharedCache.dateFormatter;
+    NSDateFormatter *f = [CKCache.sharedCache dateFormatterWithFormat:@"d"];
     f.calendar = calendar;
     f.locale = calendar.locale;
-    [f  setLocalizedDateFormatFromTemplate:@"d"];
     return [f stringFromDate:self];
 }
 
 - (NSString *)monthAndDayAndYearOnCalendar:(NSCalendar *)calendar
 {
-    NSDateFormatter *f = CKCache.sharedCache.dateFormatter;
+    NSDateFormatter *f = [CKCache.sharedCache dateFormatterWithFormat:@"MMM d yyyy"];
     f.calendar = calendar;
     f.locale = calendar.locale;
-    [f  setLocalizedDateFormatFromTemplate:@"MMM d yyyy"];
     return [f stringFromDate:self];
 }
 
 
 - (NSString *)dayOfMonthAndYearOnCalendar:(NSCalendar *)calendar
 {
-    NSDateFormatter *f = CKCache.sharedCache.dateFormatter;
+    NSDateFormatter *f = [CKCache.sharedCache dateFormatterWithFormat:@"d yyyy"];
     f.calendar = calendar;
     f.locale = calendar.locale;
-    [f  setLocalizedDateFormatFromTemplate:@"d yyyy"];
     return [f stringFromDate:self];
 }