+++ MBCalendarKit/CalendarKit/Core/CKCalendarCellContext.h
@@ -20,9 +20,9 @@ NS_SWIFT_NAME(CalendarCellContext)
 // MARK: - The Date The Cell Represents
 
 /**
- The date that the cell represents.
+ The date that the context represents.
  */
-@property (nonatomic, strong, nonnull) NSDate *date;
+@property (nonatomic, strong, nonnull, readonly) NSDate *date;
 
 // MARK: - The State that the Cell Is In
 
@@ -48,7 +48,6 @@ NS_SWIFT_NAME(CalendarCellContext)
  */
 @property (nonatomic, assign, readonly) BOOL isInSameMonthAsToday;
 
-
 /**
  `YES` if the date represented by the cell represents a date earlier than the date represented by the calendar view's `minimumDate` property.
  */
@@ -64,14 +63,14 @@ NS_SWIFT_NAME(CalendarCellContext)
 
 /**
  Create a context object. The `calendarView` is not retained.
-
+ 
  @param date The date that we are contextualizing relative to the calendar view's current state.
- @param calendarView The calendar view to use to calculate the context.
+ @param calendarView The calendar view to calculate the context for.
  @return The context object based on the current date and the configuration of the calendar view.
  */
-- (nonnull instancetype)initWithDate:(nonnull NSDate *)date andCalendarView:(nonnull CKCalendarView *)calendarView NS_DESIGNATED_INITIALIZER;
+- (nonnull instancetype)initWithDate:(nonnull NSDate *)date andCalendarView:(nonnull CKCalendarView *)calendarView;
 
-
+// Mark `init` as unavailable.
 - (nonnull instancetype)init NS_UNAVAILABLE;
 
 @end