+++ MBCalendarKit/CalendarKit/Core/CKCalendarModel+GridViewAnimationSupport.m
@@ -54,17 +54,11 @@ const void *kAnimatesWeekTransitionsKey = "com.mosheberman.calendarkit.animates-
  */
 - (BOOL)shouldAnimateTransitionFromDate:(NSDate *)fromDate toDate:(NSDate *)toDate;
 {
-    BOOL isAppropriate = YES;
-    BOOL sameMonth = [self.calendar isDate:fromDate equalToDate:toDate toUnitGranularity:NSCalendarUnitMonth];
-    BOOL sameWeek = [self.calendar isDate:fromDate equalToDate:toDate toUnitGranularity:NSCalendarUnitWeekOfYear];
+    BOOL isAppropriate = ![self isDateInSameScopeAsVisibleDateForActiveDisplayMode:fromDate];
     
-    if (self.displayMode == CKCalendarViewDisplayModeMonth && sameMonth)
+    if (self.displayMode == CKCalendarViewDisplayModeWeek)
     {
-        isAppropriate = NO;
-    }
-    if (self.displayMode == CKCalendarViewDisplayModeWeek && (sameWeek || !self.animatesWeekTransitions))
-    {
-        isAppropriate = NO;
+        isAppropriate = isAppropriate && self.animatesWeekTransitions;
     }
     
     return isAppropriate;
@@ -109,6 +103,40 @@ const void *kAnimatesWeekTransitionsKey = "com.mosheberman.calendarkit.animates-
     }
     
     return numberOfRows;
+}
+
+// MARK: - Calculating Animation Offset
+
+/**
+ Computes and returns the offset we want to start cells at.
+ We pass X and Y back to allow us to avoid checking the axis
+ in the flow layout.
+
+ @param date The final date being displayed at the end of the animation.
+ @param direction Are we moving forward or backward in time?
+ @param contentSize The size of the display.
+ @return The offset from the final position to use for cell animation.
+ */
+- (CGPoint)initialOffsetForTargetDate:(NSDate *)date forDirection:(CKCalendarTransitionDirection)direction inContentSize:(CGSize)contentSize;
+{
+    CGPoint computedSize = CGPointZero;
+    
+    if (self.transitionAxis == CKCalendarGridTransitionAxisVertical)
+    {
+        computedSize.y = contentSize.height;
+    }
+    else /* Horizontal */
+    {
+        computedSize.x = contentSize.width;
+    }
+    
+    if (direction == CKCalendarTransitionDirectionForward)
+    {
+        computedSize.x = -computedSize.x;
+        computedSize.y = -computedSize.y;
+    }
+    
+    return computedSize;
 }
 
 @end