+++ MBCalendarKit/CalendarKit/Core/CKCalendarModel.m
@@ -10,12 +10,47 @@
 #import "NSCalendar+Juncture.h"
 #import "NSCalendarCategories.h"
 
+#import "CKCalendarCellContextCache.h"
+
 @interface CKCalendarModel ()
 
 /**
  The date that was last selected by the user, either by tapping on a cell or one of the arrows in the header.
  */
 @property (nonatomic, strong) NSDate *previousDate;
+
+// MARK: - Caching Cell Context
+
+/**
+ A cache which maintains the cell contexts and responds to date, calendar, and other changes.
+ */
+//@property (nonnull, nonatomic, strong) CKCalendarCellContextCache *cellContextCache;
+
+
+// MARK: - First / Last Visible Date
+
+/**
+ The most recently cached firstVisibleDate.
+ */
+@property (nonatomic, nonnull, strong) NSDate *firstVisibleDate;
+
+
+/**
+ The most recently cached lastVisibleDate.
+ */
+@property (nonatomic, nonnull, strong) NSDate *lastVisibleDate;
+
+// MARK: - Index Paths
+
+/**
+ Cache of dates corresponding to index paths.
+ */
+@property (nonatomic, strong, nonnull) NSMutableDictionary<NSIndexPath *, NSDate *> *dateByIndexPath;
+
+/**
+ Cache of index paths corresponding to dates.
+ */
+@property (nonatomic, strong, nonnull) NSMutableDictionary<NSDate *, NSIndexPath *> *indexPathByDate;
 
 @end
 
@@ -30,8 +65,29 @@
         _calendar = [NSCalendar autoupdatingCurrentCalendar];
         _displayMode = CKCalendarViewDisplayModeMonth;
         _date = [NSDate date];
+        _firstVisibleDate = [self computedFirstVisibleDate];
+        _lastVisibleDate = [self computedLastVisibleDate];
+        
+        _indexPathByDate = [[NSMutableDictionary alloc] init];
+        _dateByIndexPath = [[NSMutableDictionary alloc] init];
     }
     return self;
+}
+
+// MARK: - Setting the Calendar
+
+- (void)setCalendar:(NSCalendar *)calendar
+{
+    _calendar = calendar;
+    [self updateVisibleDates];
+    [self informObserverOfInvalidatedState];
+}
+
+- (void)setFirstWeekday:(NSUInteger)firstWeekday
+{
+    self.calendar.firstWeekday = firstWeekday;
+    [self updateVisibleDates];
+    [self informObserverOfInvalidatedState];
 }
 
 // MARK: - Settings the Date
@@ -40,6 +96,11 @@
 {
     if (!date) {
         date = [NSDate date];
+    }
+    
+    if (date.timeIntervalSinceReferenceDate == _date.timeIntervalSinceReferenceDate)
+    {
+        return;
     }
     
     NSDateComponents *components = [self.calendar components:NSCalendarUnitYear | NSCalendarUnitMonth | NSCalendarUnitDay fromDate:date];
@@ -58,6 +119,11 @@
         }
     }
     
+    if ([self.calendar isDate:date inSameDayAsDate:_date])
+    {
+        return;
+    }
+    
     if ([self.observer respondsToSelector:@selector(calendarModel:willChangeFromDate:toNewDate:)])
     {
         [self.observer calendarModel:self willChangeFromDate:self.date toNewDate:date];
@@ -66,17 +132,51 @@
     _previousDate = self.date;
     _date = date;
     
+    if([self crossesScopeBoundaryWhenTransitioningFromDate:_previousDate toDate:_date])
+    {
+        [self updateVisibleDates];
+        //    [self.cellContextCache handleChangeSelectedDateToDate:_date];
+    }
+    
+    
     if([self.observer respondsToSelector:@selector(calendarModel:didChangeFromDate:toNewDate:)])
     {
         [self.observer calendarModel:self didChangeFromDate:self.previousDate toNewDate:self.date];
     }
+    
 }
 
 - (void)setDisplayMode:(CKCalendarViewDisplayMode)mode
 {
+    CKCalendarViewDisplayMode oldMode = _displayMode;
+    
+    if([self.observer respondsToSelector:@selector(calendarModel:willChangeFromDisplayMode:toDisplayMode:)])
+    {
+        [self.observer calendarModel:self willChangeFromDisplayMode:oldMode toDisplayMode:mode];
+    }
+    
     _displayMode = mode;
     
-    [self informObserverOfInvalidatedState];
+    [self updateVisibleDates];
+    
+    if([self.observer respondsToSelector:@selector(calendarModel:didChangeFromDisplayMode:toDisplayMode:)])
+    {
+        [self.observer calendarModel:self didChangeFromDisplayMode:oldMode toDisplayMode:mode];
+    }
+}
+
+// MARK: - Updating Visible Dates in Response to Changes
+
+
+/**
+ This method updates the visible dates and purges the caches.
+ */
+- (void)updateVisibleDates
+{
+    _firstVisibleDate = [self computedFirstVisibleDate];
+    _lastVisibleDate = [self computedFirstVisibleDate];
+    
+    [self purgeDateAndIndexPathCaches];
 }
 
 // MARK: - Getting the First Visible Date
@@ -86,7 +186,7 @@
  
  @return A date representing the first day of the week, respecting the calendar's start date.
  */
-- (nonnull NSDate *)firstVisibleDate;
+- (nonnull NSDate *)computedFirstVisibleDate;
 {
     CKCalendarViewDisplayMode displayMode = self.displayMode;
     NSDate *firstVisibleDate = self.date; /* Default to self.date */
@@ -115,7 +215,7 @@
  
  @return A date representing the first day of the week, respecting the calendar's start date.
  */
-- (nonnull NSDate *)lastVisibleDate;
+- (nonnull NSDate *)computedLastVisibleDate;
 {
     CKCalendarViewDisplayMode displayMode = self.displayMode;
     NSDate *lastVisibleDate = self.date; /* Default to self.date */
@@ -169,7 +269,6 @@
     }
     
     _maximumDate = maximumDate;
-    
     if(maximumDate && [self _dateIsAfterMaximumDate:self.date])
     {
         self.date = maximumDate;
@@ -205,7 +304,6 @@
 {
     return [self.calendar date:date isAfterDate:self.maximumDate];
 }
-
 
 /**
  Determines if the supplied date is within range of the minimum and maximum dates. 
@@ -253,5 +351,87 @@
     }
 }
 
+// MARK: - Vending Cell Contexts
+
+//- (CKCalendarCellContextCache *)cellContextCache
+//{
+//    if(!_cellContextCache)
+//    {
+//        _cellContextCache = [[CKCalendarCellContextCache alloc] initWithCalendarModel:self];
+//    }
+//    
+//    return _cellContextCache;
+//}
+
+/**
+ Returns a context object for the supplied date.
+ 
+ @param date The date for which we want context.
+ @return The context object describing the date.
+ */
+//- (nonnull CKCalendarCellContext *)contextForDate:(nonnull NSDate *)date;
+//{
+//    return  [self.cellContextCache contextForDate:date];
+//}
+
+// MARK: - Checking If Date Changes Cross Scope Boundary
+
+
+/**
+ Determine if changing from one date to another crosses scope boundary. Scope is "month" or "week" (or "day.")
+
+ @param fromDate The date before changing.
+ @param toDate The date after changing.
+ @return `YES` or `NO`.
+ */
+- (BOOL)crossesScopeBoundaryWhenTransitioningFromDate:(nonnull NSDate *)fromDate toDate:(nonnull NSDate *)toDate;
+{
+    BOOL crosses = YES;
+    
+    if (self.displayMode == CKCalendarViewDisplayModeMonth)
+    {
+        crosses = ![self.calendar isDate:fromDate equalToDate:toDate toUnitGranularity:NSCalendarUnitMonth];
+    }
+    if (self.displayMode == CKCalendarViewDisplayModeWeek)
+    {
+        crosses = ![self.calendar isDate:fromDate equalToDate:toDate toUnitGranularity:NSCalendarUnitWeekOfYear];
+    }
+    
+    return crosses;
+}
+
+/**
+ Determines if some date is in the same scope as the visible date.
+
+ @param date The date to compare.
+ @return `YES` if the date is equal to self.date to thw appropriate granularity.
+ */
+- (BOOL)isDateInSameScopeAsVisibleDateForActiveDisplayMode:(NSDate *)date;
+{
+    BOOL sameScope = NO;
+    
+    if (_displayMode == CKCalendarViewDisplayModeDay)
+    {
+        sameScope = [_calendar isDate:date inSameDayAsDate:_date];
+    }
+    else if (_displayMode == CKCalendarViewDisplayModeWeek)
+    {
+        sameScope = [_calendar isDate:date equalToDate:_date toUnitGranularity:NSCalendarUnitWeekOfYear];
+    }
+    else if(_displayMode == CKCalendarViewDisplayModeMonth)
+    {
+        sameScope = [_calendar isDate:date equalToDate:_date toUnitGranularity:NSCalendarUnitMonth];
+    }
+    
+    return sameScope;
+}
+
+// MARK: - Purge Caches
+
+- (void)purgeDateAndIndexPathCaches
+{
+    [self.dateByIndexPath removeAllObjects];
+    [self.indexPathByDate removeAllObjects];
+}
 
 @end