+++ MBCalendarKit/CalendarKit/Core/CKCalendarView+DefaultCellProviderImplementation.m
@@ -56,9 +56,8 @@
     calendarCell.state = context.identifier;
     
     /* Show the day of the month in the cell. */
-    
-    NSUInteger day = [self.calendar daysInDate:context.date];
-    calendarCell.number = @(day);
+    NSInteger day = [self.calendar component:NSCalendarUnitDay fromDate:context.date];
+    calendarCell.number = day;
     
     if([self.dataSource respondsToSelector:@selector(calendarView:eventsForDate:)])
     {